+++ Sources/BraintreeUIKit/Public/BraintreeUIKit/BTUIKExpiryFormField.h
@@ -1,14 +1,11 @@
 #ifdef COCOAPODS
-#import <BraintreeDropIn/BTUIKExpiryInputView.h>
 #import <BraintreeDropIn/BTUIKFormField.h>
 #else
-#import <BraintreeUIKit/BTUIKExpiryInputView.h>
 #import <BraintreeUIKit/BTUIKFormField.h>
 #endif
 
 /// @class Form field to collect an expiration date.
-@interface BTUIKExpiryFormField : BTUIKFormField <BTUIKExpiryInputViewDelegate>
-
+@interface BTUIKExpiryFormField : BTUIKFormField
 /// The expiration month
 @property (nonatomic, strong, nullable, readonly) NSString *expirationMonth;