+++ CHANGELOG.md
@@ -16,6 +16,8 @@
   * Enable dynamic type by default. This can be disabled with the `disableDynamicType` property on `BTDropInUICustomization`.
   * Show last four digits of vaulted cards, instead of "••• ••11", which truncates in larger fonts.
   * Remove support for Turkish localization
+  * Use number pad for expiration date input on card form
+  * Require two-digit year for expiration date instead of four-digit year
 
 ## unreleased